2024/Advent.Tests/Commands/Day17CommandTests.cs
using Advent.Common.Settings;
using Spectre.Console.Cli;

namespace Advent.Tests.Commands;

public class Day17CommandTests
{
    private readonly List<string> _arguments = [];
    private readonly IRemainingArguments _remaining = Substitute.For<IRemainingArguments>();
    private readonly TestConsole console = new();

    public Day17CommandTests()
    {
        console.Profile.Capabilities.Interactive = true;
    }

    [Fact]
    public async Task Day17Command_Solves_Part1_Correctly()
    {
        var mockReader = Substitute.For<IFileReader>();
        mockReader
            .ReadInputAsync(Arg.Any<string>())
            .Returns(Task.FromResult(TestData.Day17TestData));

        var command = new Day17Command(mockReader, console);
        var result = await command.ExecuteAsync(
            new CommandContext(_arguments, _remaining, "day17", null),
            new AdventSettings { Part = "Part 1" }
        );
        result.Should().Be(0);
        console.Output.Should().Contain("Day 17 Part 1");
        console.Output.Should().Contain("The answer is 4,6,3,5,6,3,5,2,1,0");
    }

    [Fact]
    public async Task Day17Command_Solves_Part2_Correctly()
    {
        string data =
            "Register A: 2024\n"
            + "Register B: 0\n"
            + "Register C: 0\n"
            + "\n"
            + "Program: 0,3,5,4,3,0\n";

        var mockReader = Substitute.For<IFileReader>();
        mockReader.ReadInputAsync(Arg.Any<string>()).Returns(Task.FromResult(data));

        var command = new Day17Command(mockReader, console);
        var result = await command.ExecuteAsync(
            new CommandContext(_arguments, _remaining, "day17", null),
            new AdventSettings { Part = "Part 2" }
        );
        result.Should().Be(0);
        console.Output.Should().Contain("Day 17 Part 2");
        console.Output.Should().Contain("The answer is 117440");
    }
}
